[Публичный тренажер] ANDR-72: Четвертый этап InterviewQuizScreen. Навигация + Интеграция (#150)
* ANDR-72: настройка DI, исправление минорных ошибок, добавление QuizAnswersWrapperDto

* ANDR-72: удаление лишней строки

* ANDR-72: удаление лишнего импорта

+++ core/network-api/src/main/java/ru/yeahub/network_api/models/QuestionAnswerDto.kt
@@ -1,7 +1,7 @@
 package ru.yeahub.network_api.models
 
 data class QuestionAnswerDto(
-    val questionId: Int,
+    val questionId: Long,
     val questionTitle: String,
     val answer: String
 )
